requirements.txt
Django==1.6.5
MySQL-python==1.2.5
South==1.0
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.18
celery==3.1.13
django-celery==3.1.10
django-debug-toolbar==1.2.1
django-extensions==1.3.8
django-grappelli==2.5.3
django-widget-tweaks==1.3
djangorestframework==2.3.14
gunicorn==19.0.0
kombu==3.0.21
mwclient==0.6.5
pyjade==2.2.0
pytz==2014.4
simplejson==3.5.3
six==1.7.3
sqlparse==0.1.11
wsgiref==0.1.2
